Updated all Spire projects to use new build configuration.

+++ Applications/spire/source/time_and_sales_ui_tester/periodic_time_and_sales_model.cpp
@@ -56,7 +56,7 @@ void PeriodicTimeAndSalesModel::set_period(time_duration p) {
   m_period = p;
   m_timer.stop();
   if(m_period != pos_infin) {
-    m_timer.start(m_period.total_milliseconds());
+    m_timer.start(static_cast<int>(m_period.total_milliseconds()));
   }
 }
